requirements.txt
lxml>=3.7.2
cssselect>=1.0.0
Pillow>=4.3
requests>=2.14
pycryptodome>=3.5
cloudscraper>=1.1.1
progressbar2>3.34
urllib3
packaging>=17
html-purifier>=0.1.9
js2py>=0.60
peewee>3.4.0
tinycss>=0.4
better_exceptions>=0.2
argcomplete>=1.9.4
tabulate>=0.8
